factor: changed labels, and button t removed
